deepfake-detector/models/image_detector.py
"""
Image Deepfake Detection Module
Uses Face X-ray or XceptionNet for detecting manipulated images
"""

import os
import logging
from typing import Optional
import numpy as np

logger = logging.getLogger(__name__)

class ImageDetector:
    """Image deepfake detection using pre-trained models"""
    
    def __init__(self, model_path: Optional[str] = None):
        self.model = None
        self.model_path = model_path
        self.is_loaded = False
        
    async def predict(self, image_path: str) -> float:
        """
        Predict deepfake probability for an image
        
        Args:
            image_path: Path to the image file
            
        Returns:
            float: Deepfake probability (0.0 = authentic, 1.0 = deepfake)
        """
        if not self.is_loaded:
            logger.warning("Model not loaded, using mock prediction")
            return 0.15  # Mock result for testing
            
        try:
            # Preprocess image
            processed_image = self._preprocess_image(image_path)
            
            # Run inference
            prediction = self._run_inference(processed_image)
            
            return float(prediction)
            
        except Exception as e:
            logger.error(f"Image prediction error: {e}")
            return 0.5  # Default to suspicious on error
    
    def _preprocess_image(self, image_path: str) -> np.ndarray:
        """Preprocess image for model input"""
        try:
            import cv2
            from PIL import Image
            
            # Load image
            if image_path.lower().endswith(('.png', '.jpg', '.jpeg', '.webp')):
                # Use PIL for better format support
                image = Image.open(image_path).convert('RGB')
                image = np.array(image)
            else:
                # Use OpenCV as fallback
                image = cv2.imread(image_path)
                image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            
            # Resize to model input size (typically 224x224 or 299x299)
            target_size = (224, 224)
            image = cv2.resize(image, target_size)
            
            # Normalize pixel values
            image = image.astype(np.float32) / 255.0
            
            # Add batch dimension
            image = np.expand_dims(image, axis=0)
            
            return image
            
        except Exception as e:
            logger.error(f"Image preprocessing error: {e}")
            raise
    
    def _run_inference(self, processed_image: np.ndarray) -> float:
        """Run model inference on preprocessed image"""
        try:
            if self.model is None:
                raise ValueError("Model not loaded")
            
            # This is a placeholder - replace with actual model inference
            # For example, with TensorFlow/Keras:
            # prediction = self.model.predict(processed_image)
            # return prediction[0][0]  # Assuming binary classification
            
            # For PyTorch:
            # with torch.no_grad():
            #     prediction = self.model(processed_image)
            #     return torch.sigmoid(prediction).item()
            
            # Mock inference for now
            return 0.15
            
        except Exception as e:
            logger.error(f"Inference error: {e}")
            raise

def load_image_model(model_path: Optional[str] = None) -> ImageDetector:
    """
    Load the image detection model
    
    Args:
        model_path: Optional path to model weights
        
    Returns:
        ImageDetector instance
    """
    try:
        detector = ImageDetector(model_path)
        
        # Try to load different model types
        if _try_load_face_xray(detector, model_path):
            logger.info("Face X-ray model loaded successfully")
        elif _try_load_xception(detector, model_path):
            logger.info("XceptionNet model loaded successfully")
        else:
            logger.warning("No pre-trained models found, using mock detector")
        
        return detector
        
    except Exception as e:
        logger.error(f"Failed to load image model: {e}")
        # Return mock detector for fallback
        return ImageDetector()

def _try_load_face_xray(detector: ImageDetector, model_path: Optional[str]) -> bool:
    """Try to load Face X-ray model"""
    try:
        # Face X-ray specific loading logic
        # This would typically involve:
        # 1. Loading pre-trained weights
        # 2. Setting up the model architecture
        # 3. Configuring preprocessing
        
        # For now, just mark as loaded
        detector.is_loaded = True
        return True
        
    except Exception as e:
        logger.debug(f"Face X-ray loading failed: {e}")
        return False

def _try_load_xception(detector: ImageDetector, model_path: Optional[str]) -> bool:
    """Try to load XceptionNet model"""
    try:
        # XceptionNet specific loading logic
        # This would typically involve:
        # 1. Loading pre-trained weights
        # 2. Setting up the model architecture
        # 3. Configuring preprocessing
        
        # For now, just mark as loaded
        detector.is_loaded = True
        return True
        
    except Exception as e:
        logger.debug(f"XceptionNet loading failed: {e}")
        return False